refactor: extract PlayerScreen component from App.tsx
- Create separate PlayerScreen component in components/PlayerScreen.tsx
- Move all player-related logic, state, and UI from App.tsx to PlayerScreen
- Integrate back button functionality using AppHeader component
- Simplify App.tsx to focus only on navigation routing
- Improve code organization and maintainability
